requirements.txt
# requirements.txt

python-dotenv
python-magic
colorlog
requests
requests-toolbelt
KalturaApiClient